requirements.txt
Django>=1.6,<1.9
djangorestframework>=2.3.8,<3.0
beautifulsoup4>=4.2.1
django-filter>=0.7
django_mobile>=0.3.0
pyzmq>=13.0.2
django-tagging>=0.3.1
git+https://github.com/izderadicka/rest2backbone.git#egg=rest2backbone
django-registration>=1.0.0
django-simple-captcha>=0.4.0
psycopg2>=2.4
psycogreen>=1.0
django-lockout>=0.1.1
lxml>=3.1.0
gevent-socketio>=0.3.5-rc2
geopy==0.97
six
numpy
matplotlib


